requirements.txt
numpy
scipy
pandas
joblib
scikit_learn
fire
umap_learn
matplotlib
termcolor
tqdm
